methods implemented

+++ app/src/main/java/ba/etf/rma23/projekat/UserRating.kt
@@ -3,6 +3,6 @@ package ba.etf.rma23.projekat
 data class UserRating(
     override val username: String,
     override val timestamp: Long,
-    val rating: Double
+    val rating: Int
 ): UserImpression()
